# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/bookings/me": {
      "get": {
        "operationId": "BookingController_getMyBookings",
        "summary": "Get current user bookings",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Booking"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/v1/bookings/{id}": {
      "get": {
        "operationId": "BookingController_getById",
        "summary": "Get booking by ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Booking"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/v1/bookings/{id}/cancel": {
      "post": {
        "operationId": "BookingController_cancelBooking",
        "summary": "Cancel booking",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Booking"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/bookings/{id}/check-in": {
      "post": {
        "operationId": "BookingController_checkIn",
        "summary": "Check in booking (staff+)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Booking"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/v1/temples/{templeId}/accommodation/rooms": {
      "get": {
        "operationId": "AccommodationController_listRooms",
        "summary": "List rooms for a temple",
        "parameters": [
          {
            "name": "templeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Accommodation"
        ]
      },
      "post": {
        "operationId": "AccommodationController_createRoom",
        "summary": "Create room (manager+)",
        "parameters": [
          {
            "name": "templeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Accommodation"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/temples/{templeId}/accommodation/rooms/{id}": {
      "get": {
        "operationId": "AccommodationController_getRoom",
        "summary": "Get room by ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Accommodation"
        ]
      },
      "put": {
        "operationId": "AccommodationController_updateRoom",
        "summary": "Update room (manager+)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Accommodation"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "AccommodationController_deleteRoom",
        "summary": "Delete room (admin only)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Accommodation"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/temples/{templeId}/accommodation/availability": {
      "get": {
        "operationId": "AccommodationController_getAvailability",
        "summary": "Get room availability for date range",
        "parameters": [
          {
            "name": "templeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "checkIn",
            "required": true,
            "in": "query",
            "description": "YYYY-MM-DD",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "checkOut",
            "required": true,
            "in": "query",
            "description": "YYYY-MM-DD",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Accommodation"
        ]
      }
    },
    "/api/v1/temples/{templeId}/accommodation/book": {
      "post": {
        "operationId": "AccommodationController_createBooking",
        "summary": "Create accommodation booking (hold room)",
        "parameters": [
          {
            "name": "templeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Accommodation"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/temples/{templeId}/accommodation/bookings/me": {
      "get": {
        "operationId": "AccommodationController_getMyBookings",
        "summary": "Get current user accommodation bookings",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Accommodation"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/temples/{templeId}/accommodation/bookings/{id}": {
      "get": {
        "operationId": "AccommodationController_getBooking",
        "summary": "Get booking by ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Accommodation"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/temples/{templeId}/accommodation/bookings": {
      "get": {
        "operationId": "AccommodationController_getTempleBookings",
        "summary": "Get temple bookings (staff+)",
        "parameters": [
          {
            "name": "templeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Accommodation"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/temples/{templeId}/accommodation/bookings/{id}/check-in": {
      "post": {
        "operationId": "AccommodationController_checkIn",
        "summary": "Check in booking (staff+)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Accommodation"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/temples/{templeId}/accommodation/bookings/{id}/check-out": {
      "post": {
        "operationId": "AccommodationController_checkOut",
        "summary": "Check out booking (staff+)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Accommodation"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/temples/{templeId}/accommodation/bookings/{id}/cancel": {
      "post": {
        "operationId": "AccommodationController_cancelBooking",
        "summary": "Cancel booking",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Accommodation"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Temple Digital Platform API",
    "description": "Comprehensive RESTful Backend API for Temple Digital Platform (Devotee & Admin Panel)",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}